+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased](https://github.com/kafji/shub/tree/master)
 
+## [0.3.0](https://github.com/kafji/shub/tree/v0.3.0) - 2021-12-22
+
 Breaking changes:
 
 - Rename `stars` command to `starred`